Refactor cyclist types
Getting rid of RaceConfig, which was hardly used, too. Initializing
a Cyclist is now a lot simpler than it was.

The types of cyclist have their own classes, where the definition of
their initial deck is stored, and eventually other abilities can be
defined as well. Some meta-programming puts them all into
Cyclist.TYPES, ready for an end user to choose.

Also added ability to give a Cyclist additional cards, along with
the default cards for their type.